made improvements to autostar and created change former dropdown

+++ resmap_tools/ledger_ops.py
@@ -8,6 +8,11 @@ from resmap_tools.nav_to_ledger import NavToLedgerMaster
 class LedgerScrape:
     def __init__(self, webdriver):
         self.webdriver = webdriver
+        self.base_xpath = "/html/body/table[2]/tbody/tr[4]/td/table/tbody/tr/td/table"
+        self.choose_table = {
+            "current": f"{self.base_xpath}[last()-4]/tbody/tr[2]/td/table/tbody",
+            "previous": f"{self.base_xpath}[last()-5]/tbody/tr[2]/td/table/tbody",
+        }
 
     def get_URL(self):
         return self.webdriver.driver.current_url
@@ -21,20 +26,27 @@ class LedgerScrape:
     def scrape_unit(self):
         return self.webdriver.find_element(
             By.XPATH,
-            "/html/body/table[2]/tbody/tr[4]/td/table/tbody/tr/td/table[3]/tbody/tr[2]/td/table/tbody/tr[3]/td[2]/a",
+            f"{self.base_xpath}[3]/tbody/tr[2]/td/table/tbody/tr[3]/td[2]/a",
         )
+
+    def scrape_dates(self):
+        try:
+            dates_element = self.webdriver.find_element(
+                By.XPATH,
+                f"{self.base_xpath}[3]/tbody/tr[2]/td/table/tbody/tr[5]/td[2]",
+            )
+            return dates_element.text.strip()
+        except:
+            return ""
 
     def scrape_prepaid(self):
         prepaid_element = self.webdriver.find_element(
             By.XPATH,
-            "/html/body/table[2]/tbody/tr[4]/td/table/tbody/tr/td/table[3]/tbody/tr[2]/td/table/tbody/tr[3]/td[5]",
+            f"{self.base_xpath}[3]/tbody/tr[2]/td/table/tbody/tr[3]/td[5]",
         )
         value = prepaid_element.text.strip()
-        prepaid_is_credit = self.is_credit(value) == False
+        prepaid_is_credit = self.is_payment(value) == False
         return self.webdriver.get_number_from_inner_html(value), prepaid_is_credit
-
-    def choose_table(self, num):
-        return f"/html/body/table[2]/tbody/tr[4]/td/table/tbody/tr/td/table[last(){num}]/tbody/tr[2]/td/table/tbody"
 
 
 class LedgerFunctions(LedgerScrape):
@@ -44,7 +56,7 @@ class LedgerFunctions(LedgerScrape):
         self.credit_ops = CreditMaster(self.webdriver)
         self.nav_to_ledger = NavToLedgerMaster(self.webdriver)
 
-    def is_credit(self, text):
+    def is_payment(self, text):
         return "(" in text or ")" in text
 
     def is_metered(self, row):
@@ -53,10 +65,6 @@ class LedgerFunctions(LedgerScrape):
             return True
         except:
             return False
-
-    def retrieve_rows(self, table_num):
-        table = self.webdriver.define_table(By.XPATH, self.choose_table(table_num))
-        return self.webdriver.get_rows(table)
 
     def change_ledger(self, chosen_item):
         URL = self.get_URL()
@@ -78,52 +86,77 @@ class LedgerMaster(LedgerFunctions):
 
     def delete_charge(self, transaction_element):
         self.webdriver.click_element(transaction_element)
+        if self.cancelled_true():
+            return
         self.transaction_ops.delete_charge()
 
     def credit_all_charges(self, chosen_item):
         self.webdriver.click_element(self.webdriver.return_last_element("Add Credit"))
+        if self.cancelled_true():
+            return
         self.credit_ops.credit_charge(chosen_item)
 
     def auto_allocate(self, transaction_element):
         self.webdriver.click_element(transaction_element)
+        if self.cancelled_true():
+            return
         self.transaction_ops.auto_allocate()
 
-    def allocate_charge(self, transaction_element, URL, prepaid):
+    def allocate_credit(self, transaction_element):
         self.webdriver.click_element(transaction_element)
-        self.transaction_ops.loop("allocate_charge", URL, prepaid)
+        if self.cancelled_true():
+            return
+        self.transaction_ops.loop("allocate_credit", self.URL)
 
-    def unallocate_all(self, transaction_element, operation, URL):
+    def allocate_charge(self, transaction_element, prepaid):
         self.webdriver.click_element(transaction_element)
-        self.transaction_ops.loop(operation, URL)
+        if self.cancelled_true():
+            return
+        self.transaction_ops.loop("allocate_charge", self.URL, prepaid)
+
+    def unallocate_all(self, transaction_element, operation):
+        self.webdriver.click_element(transaction_element)
+        if self.cancelled_true():
+            return
+        try:
+            self.transaction_ops.loop(operation, self.URL)
+        except:
+            self.webdriver.driver.get(self.URL)
 
     def loop_through_table(
         self,
         operation,
         chosen_item=None,
-        chosen_month=-4,
+        chosen_month="current",
         is_autostar=False,
         reloop=False,
     ):
-        URL = self.get_URL()
-        idx = 0
+        self.URL = self.get_URL()
+        table_xpath = self.choose_table[chosen_month]
         has_late_credit = False
-
+        contains_nsf = False
         if not reloop:
-            rows = self.retrieve_rows(chosen_month)
+            rows = self.webdriver.get_rows(table_xpath)
             for row in rows:
                 if self.webdriver.skip_row(row, "th3"):
                     continue
                 transaction, amount = self.get_transaction_and_amount(row)
-                if "Late Fee Credit" in transaction:
+                if "late fee credit" in transaction.lower():
                     has_late_credit = True
-        if has_late_credit:
+                if "nsf" in transaction.lower():
+                    contains_nsf = True
+
+        if has_late_credit and chosen_item != "Late Fees":
             self.loop_through_table(
                 "delete_charges", "Late Fees", chosen_month=chosen_month, reloop=True
             )
+        if contains_nsf and is_autostar:
+            return
+
+        idx = 0
 
         while True:
-            rows = self.retrieve_rows(chosen_month)
-
+            rows = self.webdriver.get_rows(table_xpath)
             redstar_in_ledger = self.webdriver.element_exists(
                 By.XPATH, '//td//font[@color="red"]'
             )
@@ -142,51 +175,71 @@ class LedgerMaster(LedgerFunctions):
 
             transaction, amount = self.get_transaction_and_amount(row)
 
+            transaction_is_nsf = "nsf" in transaction.lower()
+            if transaction_is_nsf:
+                idx += 1
+                continue
+
             if "system" in transaction.lower():
                 transaction = "System"
+            if "rule compliance" in transaction.lower():
+                transaction = "Rule Compliance"
 
-            transaction_is_nsf = "NSF" in transaction
-            if transaction_is_nsf:
-                break
+            try:
+                transaction_element = row.find_element(
+                    By.XPATH, f".//a[contains(text(), '{transaction}')]"
+                )
+            except:
+                idx += 1
+                continue
 
-            transaction_element = row.find_element(
-                By.XPATH, f".//a[contains(text(), '{transaction}')]"
+            transaction_is_payment = self.is_payment(amount)
+            transaction_is_credit = (
+                "credit" in transaction.lower() or "concession" in transaction.lower()
             )
-
-            transaction_is_credit = self.is_credit(amount)
             transaction_is_metered = self.is_metered(row)
 
             if operation == "unallocate_all":
-                if (chosen_item == "Charges" and not transaction_is_credit) or (
-                    chosen_item == "Credits" and transaction_is_credit
-                ):
-                    self.unallocate_all(transaction_element, chosen_item, URL)
+                if (
+                    (chosen_item == "Charges" and not transaction_is_payment)
+                    or (chosen_item == "Credits" and transaction_is_payment)
+                ) and not transaction_is_credit:
+                    self.unallocate_all(transaction_element, chosen_item)
 
             if operation == "allocate_all":
-                if chosen_item == "Auto" and transaction_is_credit:
+                if (
+                    chosen_item == "Credits"
+                    and transaction_is_payment
+                    and not transaction_is_credit
+                ):
                     self.auto_allocate(transaction_element)
+                # if chosen_item == "Credits" and transaction_is_credit:
+                #     self.allocate_credit(transaction_element)
 
                 prepaid, prepaid_is_credit = self.scrape_prepaid()
                 if (
                     (chosen_item == "Charges")
-                    and (not transaction_is_credit)
+                    and (not transaction_is_payment)
                     and (prepaid != 0)
                     and (prepaid_is_credit)
                 ):
-                    self.allocate_charge(transaction_element, URL, prepaid)
+                    self.allocate_charge(transaction_element, prepaid)
 
             if operation == "credit_all_charges":
-                if transaction_is_credit:
-                    self.credit_all_charges(chosen_item)
+                if not transaction_is_payment:
+                    try:
+                        self.credit_all_charges(chosen_item)
+                    except:
+                        self.webdriver.driver.get(self.URL)
+                        break
 
             if operation == "delete_charges":
                 if chosen_item == "All" or chosen_item == "Except Metered":
-                    if transaction_is_credit or (
+                    if transaction_is_payment or (
                         chosen_item == "Except Metered" and transaction_is_metered
                     ):
                         idx += 1
                         continue
-                    idx += 1
                     self.delete_charge(transaction_element)
                     continue
                 if chosen_item == "Late Fees" and "late" in transaction.lower():
@@ -194,6 +247,11 @@ class LedgerMaster(LedgerFunctions):
                     continue
 
             idx += 1
+
+    def cancelled_true(self):
+        if self.is_cancelled():
+            self.webdriver.driver.get(self.URL)
+            return True
 
     def is_cancelled(self):
         if self.thread_helper:
